# contracts/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Quicker.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/organization/companies": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "getOrganizationCompanies",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanySummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organization"
        ],
        "summary": "Create a company; its functional currency is enabled and the fiscal year containing today is opened",
        "operationId": "postOrganizationCompanies",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveCompanyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanySummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organization/companies/{companyId}": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "getOrganizationCompaniesByCompanyId",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanySummary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Organization"
        ],
        "operationId": "putOrganizationCompaniesByCompanyId",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveCompanyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanySummary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/organization/companies/{companyId}/branches/{branchId}": {
      "put": {
        "tags": [
          "Organization"
        ],
        "operationId": "putOrganizationCompaniesByCompanyIdBranchesByBranchId",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "branchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveBranchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organization/companies/{companyId}/currencies": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "getOrganizationCompaniesByCompanyIdCurrencies",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyCurrencySummary"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Organization"
        ],
        "summary": "Enable a currency for the company with its display precision and cash rounding increment",
        "operationId": "putOrganizationCompaniesByCompanyIdCurrencies",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyCurrencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyCurrencySummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organization/companies/{companyId}/settings": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "getOrganizationCompaniesByCompanyIdSettings",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SettingSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/organization/fiscal-calendars": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "getOrganizationFiscalCalendars",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FiscalCalendarSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organization"
        ],
        "operationId": "postOrganizationFiscalCalendars",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFiscalCalendarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalCalendarSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organization/fiscal-calendars/{calendarId}": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "getOrganizationFiscalCalendarsByCalendarId",
        "parameters": [
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalCalendarSummary"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Organization"
        ],
        "operationId": "putOrganizationFiscalCalendarsByCalendarId",
        "parameters": [
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFiscalCalendarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalCalendarSummary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/organization/rate-types": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "getOrganizationRateTypes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RateTypeSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organization"
        ],
        "operationId": "postOrganizationRateTypes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRateTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateTypeSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organization/rates": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "getOrganizationRates",
        "parameters": [
          {
            "name": "rateType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fromDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RateSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organization"
        ],
        "summary": "Enter a rate (1 from = rate to) valid from a date; correcting an existing one needs a reason",
        "operationId": "postOrganizationRates",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organization/rates/{rateId}": {
      "delete": {
        "tags": [
          "Organization"
        ],
        "operationId": "deleteOrganizationRatesByRateId",
        "parameters": [
          {
            "name": "rateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyCurrencyRequest": {
        "required": [
          "currency"
        ],
        "type": "object",
        "properties": {
          "currency": {
            "type": "string"
          },
          "displayDecimals": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cashRoundingIncrement": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double",
            "default": 0
          },
          "isEnabled": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "SaveBranchRequest": {
        "required": [
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "address": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "taxRegistrations": {
            "type": [
              "null",
              "object"
            ],
            "additionalProperties": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
    }
  },
...
}